+++ ext/or-tools/vendor.rb
@@ -4,7 +4,7 @@ require "fileutils"
 require "net/http"
 require "tmpdir"
 
-version = "9.2.9972"
+version = "9.3.10497"
 
 if RbConfig::CONFIG["host_os"] =~ /darwin/i
   if RbConfig::CONFIG["host_cpu"] =~ /arm|aarch64/i
@@ -17,8 +17,8 @@ if RbConfig::CONFIG["host_os"] =~ /darwin/i
 
     MSG
   else
-    filename = "or-tools_MacOsX-12.0.1_v#{version}.tar.gz"
-    checksum = "796791a8ef84507d62e193e647cccb1c7725dae4f1474476e1777fe4a44ee3e0"
+    filename = "or-tools_MacOsX-12.2.1_v#{version}.tar.gz"
+    checksum = "33941702c59983897935eef06d91aca6c89ed9a8f5f4de3a9dfe489e97d7ca8c"
   end
 else
   # try /etc/os-release with fallback to /usr/lib/os-release
@@ -33,22 +33,19 @@ else
 
   if os == "ubuntu" && os_version == "20.04"
     filename = "or-tools_amd64_ubuntu-20.04_v#{version}.tar.gz"
-    checksum = "985e3036eaecacfc8a0258ec2ebef429240491577d4e0896d68fc076e65451ec"
+    checksum = "91c3c4565c2e337f48696a3f578193912f1abefd62bc7b69e03daf1fe4f4df88"
   elsif os == "ubuntu" && os_version == "18.04"
     filename = "or-tools_amd64_ubuntu-18.04_v#{version}.tar.gz"
-    checksum = "e36406c4fe8c111e1ace0ede9d0787ff0e98f11afd7db9cc074adfd0f55628a6"
+    checksum = "6ba5cc153417267e8f8e15f8b6390b17f22de07bacc61f3740a4172ccd56c274"
   elsif os == "debian" && os_version == "11"
     filename = "or-tools_amd64_debian-11_v#{version}.tar.gz"
-    checksum = "bd49ee916213b2140ab255414d35a28f19dff7caf87632309753d3fc553f85dd"
-  elsif os == "debian" && os_version == "10"
-    filename = "or-tools_amd64_debian-10_v#{version}.tar.gz"
-    checksum = "b152fee584f0c8228fe2ff21b74c789870ff9b7064e42ca26305c6b5653f0064"
+    checksum = "db0636bab909eabf06a7004f7572dca6fa152f3823c1365b0b7428405bf250e6"
   elsif os == "centos" && os_version == "8"
     filename = "or-tools_amd64_centos-8_v#{version}.tar.gz"
-    checksum = "66ed4bb800acf92c672f7e68acdf4ea27bbfdb17bbddc02f8326cd55a97305f6"
+    checksum = "e5649069fd7a3e8228cc18b91e265a90562c5d03a0c962b0346911aada0aedc9"
   elsif os == "centos" && os_version == "7"
     filename = "or-tools_amd64_centos-7_v#{version}.tar.gz"
-    checksum = "4a5c1b1639a2828cd7e1ba82a574ef37876557b59e8aab8b81811bb750d53035"
+    checksum = "3bffdec8c09fc1345dcbd6a553437e2894014093fafb53e50adc7d4d776bb08b"
   else
     platform =
       if Gem.win_platform?